perf(portal): decouple drive sync from upload
